Updates release workflow to use Maven Central
Migrates the release workflow to directly publish to Maven Central,
simplifying the process and leveraging the official repository.

This involves updating the GitHub Actions workflow to use the
`actions/setup-java@v4` action with appropriate Maven Central
credentials and GPG configuration, as well as modifying the `pom.xml`
to point to the correct Maven Central repositories.

Removes the nexus-staging-maven-plugin from the pom.xml as it's not needed
anymore.